server/wrc2025/requirements.txt
requests>=2.0.0
PyPDF2>=3.0.1
pdfplumber>=0.6.0
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib